trt/archiver/requirements.txt
asn1crypto==0.24.0
cryptography==2.1.4
graphsurgeon==0.4.5
idna==2.6
pycrypto==2.6.1
PyYAML==6.0
six==1.11.0
tensorrt==7.2.2.3
